Bring x86_64 self-tests to parity
Port the raw-syscall IPC, cross-LP, ping-pong, and device capability tests to x86_64. Clear the x86_64 Clippy backlog and enforce warning-clean linting for both architectures in CI.

+++ crates/catten/src/device/vt_d.rs
@@ -689,5 +689,5 @@ pub fn pending_fault_events() -> u32 {
     if base == 0 {
         return 0;
     }
-    ((read32(base, FSTS) >> 1) & 1) as u32
+    (read32(base, FSTS) >> 1) & 1
 }
